SystemUI: fix notifs not coming back after unmarking as spam
When a notification gets marked as spam, the view is removed from
SystemUI, but it is not removed from the internal
NotificationManagerService notifications list.

This is not a problem until removing the notification from the
spam filter. After that, any notifications posted try to go through
updateNotification() instead of addNotification(). SystemUI won't post
those notifications since there's not one to update.

When removing a notification from the spam filter, remove any internal
references to the notification so it can be posted new again.

We could remove the reference when the notification gets marked as spam,
but we need to eventually clear the cache *only* if it gets removed from
the spam list. The spam cache gets updated regularly on notification
posts.

Change-Id: Icecba34cf6de18d5f217c8c69ba8db5408e6db82
Signed-off-by: Roman Birg <roman@cyngn.com>